core/standards/web-apps/media/semantics/buffered/CORE-38409.html

<!doctype html>
<title>CORE-38409: Buffered timerange should contain the new currentTime that media is seeked to - webm</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/core/standards/web-apps/media/support/common.js"></script>
<p id="log">FAIL (script didn't run)</p>
<audio preload="metadata" controls></audio>
<video preload="metadata" controls></video>
<script>
var tests = init_tests("HTMLAudioElement.buffered: Buffered timerange should contain the new currentTime that media is seeked to reach - webm", "HTMLVideoElement.buffered: Buffered timerange should contain the new currentTime that media is seeked to - webm", {timeout:10000});
tests.forEach(function(vars) {
    var t = vars[0];
    t.step(function() {
    var tag_name = vars[1];
    var node = document.getElementsByTagName(tag_name)[0];
    node.addEventListener('loadedmetadata', function() {node.currentTime = 300;}, false);
    node.addEventListener('seeked', endTest, false);
    node.src = '/resources/media/webm/big_buck_bunny_480p.webm?nocache=' + Math.random();

    function endTest() {
        t.step(function() {
        assert_true(node.buffered.start(1) <= 300, 'The start of the buffered timerange should be <= 300');
        assert_true(node.buffered.end(1) >= 300, 'The end of the buffered timerange should be >= 300');
        t.done();
        });
    }
    });
});
</script>
